requirements.txt
cffi==1.7.0
cryptography==1.4
dj-database-url==0.4.1
Django==1.9.7
django-activity-stream==0.6.2
django-ajax==0.2.0
django-dajax==0.9.2
django-dajaxice==0.7
django-el-pagination==2.1.2
django-endless-pagination==2.0
django-model-utils==2.5
django-notifications-hq==1.2
djangocms-installer==0.8.11
gevent==1.1.1
-e git://github.com/abourget/gevent-socketio.git@8971328ab467adc49bec98bc9eb7d13c89d9dac6#egg=gevent_socketio
gevent-websocket==0.9.5
greenlet==0.4.10
idna==2.1
jsonfield==1.0.3
ndg-httpsclient==0.4.2
Pillow==3.3.0
pusher==1.3.0
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.0.0
pytz==2016.6.1
requests==2.10.0
six==1.10.0
tzlocal==1.2.2
urllib3==1.16
virtualenv==15.0.1
virtualenvwrapper-win==1.2.1
